+++ src/main/java/http/response/Response.java
@@ -1,4 +1,4 @@
-package http;
+package http.response;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
