+++ src/main/java/webserver/RequestHandler.java
@@ -1,20 +1,16 @@
 package webserver;
 
-import http.*;
+import http.request.Request;
+import http.response.Response;
+import http.response.ResponseHeader;
+import http.utils.RequestFactory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import utils.IOUtils;
-import webserver.controller.Controller;
-import webserver.controller.CreateUserController;
-import webserver.controller.FileController;
 import webserver.support.MethodHandler;
 
 import java.io.*;
 import java.net.Socket;
 import java.net.URISyntaxException;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
 
 public class RequestHandler implements Runnable {
     private static final Logger logger = LoggerFactory.getLogger(RequestHandler.class);
@@ -30,7 +26,8 @@ public class RequestHandler implements Runnable {
                 connection.getPort());
 
         try (InputStream in = connection.getInputStream(); OutputStream out = connection.getOutputStream()) {
-            Request request = readRequestUrl(in);
+//            Request request = readRequestUrl(in);
+            Request request = RequestFactory.makeRequest(in);
             Response response = new Response(new ResponseHeader());
 
             MethodHandler handler = new MethodHandler();
@@ -42,16 +39,16 @@ public class RequestHandler implements Runnable {
         }
     }
 
-    private Request readRequestUrl(InputStream in) throws IOException {
-        InputStreamReader inputStreamReader = new InputStreamReader(in);
-        BufferedReader br = new BufferedReader(inputStreamReader);
-        RequestHeader header = new RequestHeader(IOUtils.parseData(br));
-
-        if (header.get(RequestHeader.HTTP_METHOD).equals("POST")) {
-            String body = IOUtils.readData(br, Integer.parseInt(header.get("content-length")));
-            RequestBody requestBody = new RequestBody(body);
-            return new Request(header, requestBody);
-        }
-        return new Request(header);
-    }
+//    private Request readRequestUrl(InputStream in) throws IOException {
+//        InputStreamReader inputStreamReader = new InputStreamReader(in);
+//        BufferedReader br = new BufferedReader(inputStreamReader);
+//        RequestHeader header = new RequestHeader(IOUtils.parseData(br));
+//
+//        if (header.extractHeader(RequestHeader.HTTP_METHOD).equals("POST")) {
+//            String body = IOUtils.readData(br, Integer.parseInt(header.extractHeader("content-length")));
+//            RequestBody requestBody = new RequestBody(body);
+//            return new Request(requestLine, header, requestBody);
+//        }
+//        return new Request(header, requestLine);
+//    }
 }
