+++ src/test/java/http/RequestBodyTest.java
@@ -1,5 +1,6 @@
 package http;
 
+import http.request.RequestBody;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
